지형 높이 늘림

+++ Player.h
@@ -91,7 +91,7 @@ public:
 	void SetCamera(CCamera *pCamera) { m_pCamera = pCamera; }
 		
 	//플레이어를 이동하는 함수이다. 
-	void Move(ULONG nDirection, float fDistance, bool bVelocity = false) ;
+	void Move(DWORD nDirection, float fDistance, bool bVelocity = false) ;
 	void Move(XMFLOAT3& xmf3Shift, bool bVelocity = false);
 	void Move(float fxOffset = 0.0f, float fyOffset = 0.0f, float fzOffset = 0.0f) ;
